Create Triangle.java
In OpenGl we need to define a class for every Object i.e every image we need to draw, this is the triangle class contains draw method that draws the triangle to the surface-view. Remember in OPenGl all the images are advised to construct using the triangles. for example rectangle is drawn using two triangles.see Rectangle.java to know how its done.